visualization/Sentilytics_Real_Time_Live_Streamlit_dashboard/requirements.txt
numpy
pandas
plotly.express
streamlit
wordcloud
dotenv